+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
 -   Increase tolerance duration for superevent existence nagios check in
     the last hour from 1 min to 10 mins.
 
+-   Add error handling to hop-client consumers.
+
 2.1.4 "Mokele-Mbembe" (07-26-2023)
 ----------------------------------